week 2 programming assignment - add main() for programming assignment - update SortAndCountInversions to BigInt

+++ src/main/java/week_2/SortCountResult.java
@@ -1,9 +1,11 @@
 package week_2;
 
+import java.math.BigInteger;
+
 public class SortCountResult {
     private final Integer[] array;
-    private final int numInversions;
-    public SortCountResult(Integer[] intArray, int numInversions) {
+    private final BigInteger numInversions;
+    public SortCountResult(Integer[] intArray, BigInteger numInversions) {
         this.array = intArray;
         this.numInversions = numInversions;
     }
@@ -12,7 +14,7 @@ public class SortCountResult {
         return array;
     }
 
-    public int getNumInversions() {
+    public BigInteger getNumInversions() {
         return numInversions;
     }
 }
